Move logging-related tests to own file
