+++ tests/cli.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::bool_assert_comparison)]
 
+mod cases;
 mod server;
 
 use std::collections::{HashMap, HashSet};
@@ -121,6 +122,18 @@ const BINARY_SUPPRESSOR: &str = concat!(
     "+-----------------------------------------+\n",
     "\n"
 );
+
+#[allow(unused)]
+mod prelude {
+    pub(crate) use super::color_command;
+    pub(crate) use super::get_base_command;
+    pub(crate) use super::get_command;
+    pub(crate) use super::random_string;
+    pub(crate) use super::redirecting_command;
+    pub(crate) use super::server;
+    pub(crate) use super::RequestExt;
+    pub(crate) use super::BINARY_SUPPRESSOR;
+}
 
 #[test]
 fn basic_json_post() {
@@ -3609,60 +3622,4 @@ fn multiple_format_options_are_merged() {
 
 
         "#});
-}
-
-#[test]
-fn logs_are_printed_in_debug_mode() {
-    get_command()
-        .arg("--debug")
-        .arg("--offline")
-        .arg(":")
-        .env_remove("RUST_LOG")
-        .assert()
-        .stderr(contains("DEBUG xh] Cli {"))
-        .success();
-}
-
-#[test]
-fn logs_are_not_printed_outside_debug_mode() {
-    get_command()
-        .arg("--offline")
-        .arg(":")
-        .env_remove("RUST_LOG")
-        .assert()
-        .stderr("")
-        .success();
-}
-
-#[test]
-fn backtrace_is_printed_in_debug_mode() {
-    let mut server = server::http(|_req| async move {
-        panic!("test crash");
-    });
-    server.disable_hit_checks();
-    get_command()
-        .arg("--debug")
-        .arg(server.base_url())
-        .env_remove("RUST_BACKTRACE")
-        .env_remove("RUST_LIB_BACKTRACE")
-        .assert()
-        .stderr(contains("Stack backtrace:"))
-        .failure();
-}
-
-#[test]
-fn backtrace_is_not_printed_outside_debug_mode() {
-    let mut server = server::http(|_req| async move {
-        panic!("test crash");
-    });
-    server.disable_hit_checks();
-    let cmd = get_command()
-        .arg(server.base_url())
-        .env_remove("RUST_BACKTRACE")
-        .env_remove("RUST_LIB_BACKTRACE")
-        .assert()
-        .failure();
-    assert!(!std::str::from_utf8(&cmd.get_output().stderr)
-        .unwrap()
-        .contains("Stack backtrace:"));
 }
